requirements.txt
streamlit
supabase
google-generativeai
Pillow
pandas
requests
